prj_summer-badge/app-py_client-tcp/requirements.txt
opencv_python==4.7.0.68
Pillow==9.4.0
PyAutoGUI==0.9.54
pyttsx3==2.90
pyzbar==0.1.9